Updated application to make better use of SASS variables. Initial round on hooking into services for projects and records. Added recent projects view in side navbar. Added launch button to project cards. Added prompt for when no sources are found for a project. Added delete confirmation to project-card deletion. Updated Rails Record lookup to incorporate project id in SQL query.

+++ ui/src/app/projects/project-card/project-card.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
+import { Router } from '@angular/router';
 import { Project } from '../../shared/models/index';
-import { ProjectDaoService } from '../../shared/services/project-dao.service';
+import { ProjectDaoService } from '../../shared/services/index';
 
 @Component({
   selector: 'rvn-project-card',
@@ -13,26 +14,27 @@ export class ProjectCardComponent implements OnInit {
   @Input() editing:boolean = false;
   @Input() canEdit:boolean = true;
   @Input() canDelete:boolean = true;
+  @Input() pendingDelete:boolean = false;
   @Output() deleted:EventEmitter<Project> = new EventEmitter<Project>();
   @Output() saved:EventEmitter<Project> = new EventEmitter<Project>();
   @Output() created:EventEmitter<Project> = new EventEmitter<Project>();
 
-  constructor(private _projectDaoService:ProjectDaoService) {
-  }
+  constructor(private _projectDaoService:ProjectDaoService, private _router:Router) { }
 
   ngOnInit() {
   }
 
   save() {
     this.editing = false;
-    
-    if  (this.project.id) {
+    this.pendingDelete = false;
+
+    if (this.project.id) {
       this._projectDaoService.saveProject(this.project)
         .subscribe((project:Project) => {
           this.project = project;
           this.saved.emit(this.project);
         });
-    } else  {
+    } else {
       this._projectDaoService.createProject(this.project)
         .subscribe((project:Project) => {
           this.project = project;
@@ -43,16 +45,27 @@ export class ProjectCardComponent implements OnInit {
 
   cancel() {
     this.editing = false;
+    this.pendingDelete = false;
   }
 
   edit() {
     this.editing = true;
+    this.pendingDelete = false;
   }
 
   destroy() {
-    this._projectDaoService.deleteProject(this.project)
-      .subscribe(() => {
-        this.deleted.emit(this.project);
-      });
+    
+    if (this.pendingDelete) {
+      this._projectDaoService.deleteProject(this.project)
+        .subscribe(() => {
+          this.deleted.emit(this.project);
+        });
+    } else {
+      this.pendingDelete = true;
+    }
+  }
+  
+  launch() {
+    this._router.navigate(['project', this.project.id]);
   }
 }
